# be/docs/swagger.yaml
definitions:
...
  codon-backend_internal_models.Chapter:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      order_index:
        type: integer
      subject_id:
        type: string
      updated_at:
        type: string
    type: object
...
  codon-backend_internal_models.Subject:
    properties:
      course_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      order_index:
        type: integer
      updated_at:
        type: string
    type: object
...
  internal_handlers.createChapterRequest:
    properties:
      description:
        example: Laws of thermodynamics
        type: string
      name:
        example: Thermodynamics
        type: string
      order_index:
        example: 1
        type: integer
    type: object
...
  internal_handlers.createSubjectRequest:
    properties:
      description:
        example: The study of matter and energy
        type: string
      name:
        example: Physics
        type: string
      order_index:
        example: 1
        type: integer
    type: object
...
  internal_handlers.curriculumChapterView:
    properties:
      content_count:
        example: 5
        type: integer
      id:
        type: string
      name:
        example: Thermodynamics
        type: string
      order_index:
        example: 1
        type: integer
      test_count:
        example: 2
        type: integer
    type: object
  internal_handlers.curriculumCourseView:
    properties:
      id:
        type: string
      name:
        example: NEET UG
        type: string
      subjects:
        items:
          $ref: '#/definitions/internal_handlers.curriculumSubjectView'
        type: array
    type: object
  internal_handlers.curriculumResponse:
    properties:
      course:
        $ref: '#/definitions/internal_handlers.curriculumCourseView'
    type: object
  internal_handlers.curriculumSubjectView:
    properties:
      chapters:
        items:
          $ref: '#/definitions/internal_handlers.curriculumChapterView'
        type: array
      id:
        type: string
      name:
        example: Physics
        type: string
      order_index:
        example: 1
        type: integer
    type: object
...
  internal_handlers.errorResponse:
    properties:
      code:
        example: pool_too_small
        type: string
      details: {}
      error:
        example: descriptive error message
        type: string
    type: object
...
  internal_handlers.updateChapterRequest:
    properties:
      description:
        type: string
      name:
        type: string
      order_index:
        type: integer
    type: object
...
  internal_handlers.updateSubjectRequest:
    properties:
      description:
        type: string
      name:
        type: string
      order_index:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/admin/courses/{course_id}/subjects:
    post:
      consumes:
      - application/json
      description: Creates a new subject under a specific course.
      parameters:
      - description: Course UUID
        in: path
        name: course_id
        required: true
        type: string
      - description: Subject details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.createSubjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/codon-backend_internal_models.Subject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Create a Subject (Admin)
      tags:
      - Curriculum
...
  /api/v1/admin/subjects/{id}:
    patch:
      consumes:
      - application/json
      description: Partially updates a subject.
      parameters:
      - description: Subject UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.updateSubjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/codon-backend_internal_models.Subject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Update a Subject (Admin)
      tags:
      - Curriculum
...
  /api/v1/courses/{id}/curriculum:
    get:
      description: Returns the nested hierarchy of a course -> subjects -> chapters,
        including counts of published content and tests.
      parameters:
      - description: Course UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.curriculumResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      summary: Get course curriculum hierarchy
      tags:
      - Curriculum
...
  /api/v1/teacher/chapters/{id}:
    patch:
      consumes:
      - application/json
      description: Partially updates a chapter. Callable by teachers and admins.
      parameters:
      - description: Chapter UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.updateChapterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/codon-backend_internal_models.Chapter'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Update a Chapter (Teacher)
      tags:
      - Curriculum
...
  /api/v1/teacher/subjects/{subject_id}/chapters:
    post:
      consumes:
      - application/json
      description: Creates a new chapter (topic) under a specific subject. Callable
        by teachers and admins.
      parameters:
      - description: Subject UUID
        in: path
        name: subject_id
        required: true
        type: string
      - description: Chapter details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.createChapterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/codon-backend_internal_models.Chapter'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Create a Chapter (Teacher)
      tags:
      - Curriculum
...
swagger: "2.0"